Removed removeBlankRows in TableSetCleaner.class
Because:
-- It is marked as deprecated
-- It could introduced unexpected behavior
-- It is throwing out of bound exceptions on blank workbooks

Removed tests and other methods associated with it's functionality. The method calls on the first column
in the table. If it is absent it removes the rows. I am concerned this will result in unexpected bugs.
Rows should be removed if all values are empty. Test were removed as well.